rewrite markdown page links to markdown urls
