+++ docs/scripts/generate-llms/shared.js
@@ -2,12 +2,18 @@ import fs from 'node:fs';
 import path from 'node:path';
 
 import { buildAgentInstructions } from '../agent-instructions.ts';
+import {
+  DOCS_BASE_URL,
+  getMarkdownHref,
+  getMarkdownUrl,
+  rewriteDocsLinksToMarkdown,
+} from '../markdown-link-utils.ts';
+
+export { DOCS_BASE_URL, getMarkdownHref, getMarkdownUrl, rewriteDocsLinksToMarkdown };
 
 export const OUTPUT_DIRECTORY_NAME = 'public';
 export const BUILD_OUTPUT_DIR = 'out';
-export const DOCS_BASE_URL = 'https://docs.expo.dev';
 
-const DOCS_BASE_URL_WITH_SLASH = `${DOCS_BASE_URL}/`;
 const LLMS_FILES = [
   { filename: 'llms.txt', description: 'A list of all available documentation files' },
   {
@@ -47,101 +53,6 @@ const AGENT_INSTRUCTIONS_BLOCK_REGEX = /<AgentInstructions>[\S\s]*?<\/AgentInstr
 
 const DOC_INDEX_BLOCKQUOTE =
   '> For the complete documentation index, see [llms.txt](/llms.txt). Use this Use this file to discover all available pages.';
-
-function splitUrlSuffix(url) {
-  const suffixIndex = url.search(/[#?]/);
-  if (suffixIndex === -1) {
-    return { pathname: url, suffix: '' };
-  }
-
-  return {
-    pathname: url.slice(0, suffixIndex),
-    suffix: url.slice(suffixIndex),
-  };
-}
-
-function hasFileExtension(pathname) {
-  const match = pathname.match(/\.([^./]+)$/);
-  // A purely numeric "extension" (e.g. the trailing `0` in `/versions/v51.0.0`) is part of a
-  // version segment, not a real file extension, so such hrefs should still be converted to `.md`.
-  return match !== null && !/^\d+$/.test(match[1]);
-}
-
-export function getMarkdownHref(href) {
-  if (!href || href.startsWith('#')) {
-    return href;
-  }
-
-  const { pathname, suffix } = splitUrlSuffix(href);
-  if (!pathname || hasFileExtension(pathname)) {
-    return href;
-  }
-
-  const normalizedPathname =
-    pathname === '/' ? '/index' : pathname.replace(/\/+$/g, '').replace(/^([^/])/, '/$1');
-
-  return `${normalizedPathname}.md${suffix}`;
-}
-
-export function getMarkdownUrl(href) {
-  return `${DOCS_BASE_URL}${getMarkdownHref(href)}`;
-}
-
-// Matches the opening or closing line of a fenced code block. CommonMark allows up to three
-// leading spaces and supports both backtick and tilde fences, so we accept indented fences
-// (e.g. those nested inside list items) and `~~~` fences, not only ``` at column 0.
-const CODE_FENCE_REGEX = /^\s{0,3}(`{3,}|~{3,})/;
-
-function rewriteMarkdownLinks(line) {
-  return line.replace(
-    /(!?\[[^\n\]]+]\()([^\s)]+)((?:\s+["'][^\n)]*["'])?\))/g,
-    (match, prefix, url, suffix) => {
-      if (prefix.startsWith('!')) {
-        return match;
-      }
-
-      if (url.startsWith(DOCS_BASE_URL_WITH_SLASH)) {
-        return `${prefix}${getMarkdownUrl(url.slice(DOCS_BASE_URL.length))}${suffix}`;
-      }
-
-      if (url.startsWith('/') && !url.startsWith('//')) {
-        return `${prefix}${getMarkdownHref(url)}${suffix}`;
-      }
-
-      return match;
-    }
-  );
-}
-
-export function rewriteDocsLinksToMarkdown(content) {
-  // The fence marker (`` ` `` or `~`) that opened the current code block, or null when outside one.
-  // Tracking the marker keeps a ``` line from prematurely closing a ~~~ block and vice versa.
-  let fenceMarker = null;
-
-  return content
-    .split('\n')
-    .map(line => {
-      const fenceMatch = line.match(CODE_FENCE_REGEX);
-      if (fenceMatch) {
-        const marker = fenceMatch[1][0];
-        if (fenceMarker === null) {
-          fenceMarker = marker;
-          return line;
-        }
-        if (marker === fenceMarker) {
-          fenceMarker = null;
-          return line;
-        }
-      }
-
-      if (fenceMarker !== null) {
-        return line;
-      }
-
-      return rewriteMarkdownLinks(line);
-    })
-    .join('\n');
-}
 
 export function stripAgentInstructions(content) {
   return content.replace(AGENT_INSTRUCTIONS_BLOCK_REGEX, '');
